api auth/requirements.txt
flask==2.2.2
flask_sqlalchemy==3.0.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
python-decouple==3.6

